ft-#151289585 Restructure files to allow travis build
